JDK-8285362
